refactor(basis): move IFDD to dedicated basis submodule and remove adaptive
