--- src/basis/adaptive/mod.rs
@@ -1,2 +1,0 @@
-//! Module for _adaptive_ basis representations.
-import_all!(ifdd);
