+++ src/basis/mod.rs
@@ -1,11 +1,11 @@
 //! Module for basis function representations used by `LFA`.
 import_all!(composable);
 
-pub mod adaptive;
 pub mod composition;
 pub mod fixed;
+pub mod ifdd;
 
 pub use crate::{
-    core::{Projector, AdaptiveProjector, Projection, Feature, CandidateFeature},
+    core::{Projector, Projection},
     geometry::kernels,
 };
