+++ src/core/mod.rs
@@ -7,7 +7,6 @@ import_all!(error);
 
 import_all!(approximator);
 
-import_all!(feature);
 import_all!(projection);
 import_all!(projector with macros);
